Add contributing notes to README
README file should contain instructions about
how to work with the project.
At the moment there is nothing to do with it except
asking questions, adding documentation, basic code
and criticizing blog in commits.
CONTRIBUTING and contributing coveant already exist,
so it's nice to add links to them to README.
Now everyone, who enters the repository,
will now how to create new pull request
and that it's a bad idea to use rude language.