Fix link name to Code of Conduct in CONTRIBUTING
Found the root of `Contributor Covenant` usage instead of `Code of Conduct`.
In 2ede5e68ff56361a819ca85485f7a793f5f69eaf it was copied and pasted
from CONTRIBUTING to README file.
Change it here, because `Code of Conduct` is the name of referenced file
and its sense as well.