Add CHANGELOG file
CHANGELOG file is a very useful thing.
As a developer, sometimes I need to know,
when some features were created,
without jumping with `git blame -c` or `git log --follow`
(I could change name of some classes after a while).
As a user, I need to know whether it's a nice idea
to use a new version of some library without comparing
documentation of different versions.
Also I need a guarantee that my code will not be broken
with new release of the library.
SemVer helps with this.
If a library has a major version greater than zero,

- removal of any features or behavior change of existent ones
  should lead to bump of major version.
- new features should go with new minor version;
- bug fixes, that don't change API, change patch version.

Simple enough.
If major version is zero, then minor version plays role of major,
and patch plays role of both minor and patch version.
When only patch version is not zero,
any version changes affect it until contributors
will bump minor or major version.

Seems like this should work very nice
after the first major version release.
Though, when different parts of public API
deprecated (and then deleted) frequently,
major version of a library grows fast.
As the result, we may get `Stereo Parallel 35`,
and unlikely to support the first on even tenth version.
Well, when this will happen, I'll think about another versioning.
For example CalVer (Calendar Versioning)
https://calver.org/

I don't know whether I really need table of contents
for the CHANGELOG, but let it be.
Headers, which are specified in table of contents,
are clickable and have a backward reference to table of contents
(at leat, on GitHub).
Though, headings in CHANGELOG will refer to comparison pages
between current and previous tag.
Table of contents does not break this, so I'm happy with it.