Add instructions for version bump to CONTRIBUTING
Now we have a CHANGELOG and it's needed to support it
when someone makes changes to the project.

Tagging is a convenient mechanism to store versions of projects
written in languages that don't support versioning.
Also it's very handy to download archive with corresponding tag
from GitHub.

I think, it's a good idea to copy section from CHANGELOG
to description of corresponding tag.
I've never used the description of tag,
but it sounds better than nothing.

I propose to write project name in bump commit title
to show explicitly that the version of the main project was changed,
not its dependencies'.