Add Travis CI badge to README
Refer to this page to see how to do this
https://docs.travis-ci.com/user/status-images
Badge is a very cool thing:
when I see a repository, I may want to know, what integration it has.
With badges in the top of README I can see this information
and use that services in my projects too.
Now, visitors of my repository, who don't know about Travis CI,
can click on the badge and find out something new.