Add CMakeLists file for tests
Here an interesting part begins.
Just go here to read FAQ of Boost Test library compilation
https://www.boost.org/doc/libs/1_62_0/libs/test/doc/html/boost_test/section_faq.html
It contains an example of CMakeLists.txt file
for the shared libraries variant of the fwamework.
I haven't got an idea of setting `Boost_ADDITIONAL_VERSIONS` variable,
so I avoid this step.
Also, some commands are stored in the root CMakeLists.txt file,
so they are not needed here.

First, we need to find the package.
https://cmake.org/cmake/help/v3.0/command/find_package.html
We specify its name, minimal version, and mark it as `REQUIRED`
to fail build if the package will be not found.
Next, we write `COMPonentS` and enumerate needed components.
We don't need all Boost dependencies -- here we need only
`unit_test_framework`.
I've specified version `1.54`, because it's supported by
Ubuntu Trusty 14.04
https://packages.ubuntu.com/trusty/libboost-all-dev
Fedora 26 has `1.63`
https://fedora.pkgs.org/26/fedora-x86_64/boost-1.63.0-5.fc26.x86_64.rpm.html
Debian Jessie has `1.55`
https://packages.debian.org/ru/jessie/libboost-all-dev
Version `1.54` was released on July 1st, 2013
https://www.boost.org/users/history/

Next, create executable file with test
https://cmake.org/cmake/help/v3.0/command/add_executable.html
I call it `test_executable` like in Boost FAQ,
and use the only source file `api.cpp`.

To use Boost library, we include its headers.
Compiler needs to know where to get them
https://cmake.org/cmake/help/v3.0/command/target_include_directories.html
Name of target is a name of the test executable.
Here is a good explanation of what `PRIVATE` and `PUBLIC` means here
https://stackoverflow.com/a/28305481/1305036
If test uses the Boost library,
but another mmodules that use its headers
don't have an access to the Boost library via those headers of the test,
we write `PRIVATE`.
In other case, we should use `PUBLIC`.

Let's link shared libary of the Boost Test to the target
https://cmake.org/cmake/help/v3.0/command/target_link_libraries.html

We need to define `BOOST_TEST_DYN_LINK` macro for Boost to know
that we want to use shared library.
Also it's needed to specify name of the module
by `BOOST_TEST_MODULE` macro.
This can be done in CMakeLists.txt file too
https://cmake.org/cmake/help/v3.0/command/target_compile_definitions.html

At last, we need to add the test for `ctest` to recognize it.
https://cmake.org/cmake/help/v3.0/command/add_test.html
Looks quite simple: set name of the test and command to call to run it.