Add tests directory to main CMakeLists.txt
We want to run our tests using `ctest`
https://cmake.org/cmake/help/v3.0/manual/ctest.1.html
It's convenient, because we don't need to care about
how to launch tests based on testing framework
used by the project.
So, we just build the project and call `ctest`

```
cmake --build .
ctest
```

To enable this cool feature, `enable_testing` should be called
in the root CMakeLists.txt file
https://cmake.org/cmake/help/v3.0/command/enable_testing.html

Next, we need to add directory with tests
(and CMakeLists.txt file for them)
to the project using `add_subdirectory`
https://cmake.org/cmake/help/v3.0/command/add_subdirectory.html

It works!
Need to add this to README file and Travis CI configuration file.