Initialize CMakeLists file for documentation build
Create an option to specify
whether we want to build documentation or not.
It's simple:
call `option` function, specify variable name,
add description of its purpose and set default value
https://cmake.org/cmake/help/v3.0/command/option.html

Name `BUILD_DOC` is good enough and `Build documentation` seems clear.
By default, we don't build the documentation,
because when we build the project we want to use it.
If we want to read the documentation,
we'll rather go to website with compiled documentation.

If we want to build the documentation,
try to find Doxygen by `find_package`.
Mark Doxygen as required,
 because we cannot build the documentation when it's not available.
Finding Doxygen is described in documentation for CMake 3.9,
but available for 3.0 too,
but documentation about this is small there
https://cmake.org/cmake/help/v3.9/module/FindDoxygen.html