Add documentation build to CMake
I follow the guide posted by Victoria Rudakova
https://vicrucann.github.io/tutorials/quick-cmake-doxygen
Lot of magic I see there.

Set path to Doxygen configuration file
that is in the same directory as current CMakeLists.txt.
`CMAKE_CURRENT_SOURCE_DIR` will help us.
https://cmake.org/cmake/help/v3.0/variable/CMAKE_CURRENT_SOURCE_DIR.html
Set where to copy this file.
We want to place it into corresponding binary directory.
Use `CMAKE_CURRENT_BINARY_DIR`
https://cmake.org/cmake/help/v3.0/variable/CMAKE_CURRENT_BINARY_DIR.html
Copy input file to its destination using `configure_file`.
The command is used to substitute variable names with values.
At the moment I don't have any variables in the file,
so I use `COPYONLY` option
https://cmake.org/cmake/help/v3.0/command/configure_file.html

We are ready to make the magic.
Use `add_custom_target`
to add documentation target and build it using custom command.
Its name will be `doc_doxygen`.
In Victoria's post it's said that `ALL` option
makes the documentation to be build with all application
https://cmake.org/cmake/help/v3.0/command/add_custom_target.html
Execute `DOXYGEN_EXECUTABLE` and store result into `DOXYGEN_OUT`.
Specify current binary folder to be a `WORKING_DIRECTORY`,
as if we execute Doxygen from it.
`COMMENT` is a message to be shown before command execution.
As I understand,
`VERBATIM` option escapes symbols of parameters provided to the command
to pass them correctly and unchanged to it.
I'm not sure that that's all, but let it be there.