Add badge from CodeDocs[xyz]
I've integrated the project with the service successfully
and I want to have a nice badge that will show
that the documentation exists.

On page of the service I can see a list of my project.
Enabled projects have links to documentation
and badge with documentation status.
Just copy the link and address of the image
and add it to README file of the project.