Add Image library to CMake project
To make the Image structure available in unit tests
we can create a library for it.

Use `add_library`.
the first argument is a name of the library.
It should be unique for all project in order to have an ability
to add the library to any other library or executable.
The next arguments are files that need to be included into the library.
There is also a possibility to choose between static, shared library or
module mode.
By default, static library is used.
This provikes creation of archived object file that will be linked by
linker during project build.
This allows you to catch unresolved symbols on build stage.
If `BUILD_SHARED_LIBS` flag is enabled, the library will be built as a
shared and will be loaded at runtime.
This allows you to send updates of your project by smaller parts
without need to recompile it.
Also, there is a `MODULE` option, and I didn't get its meaning.
I leave this behavior, because I don't see a reason to write more words.
Let user decide how to compile libraries of the project
https://cmake.org/cmake/help/v3.0/command/add_library.html

The project didn't compile,
because CMake didn't know what language it should use
to link the `image.hpp` file.

```
-- The CXX compiler identification is GNU 7.3.0
-- Check for working CXX compiler: /usr/bin/c++
-- Check for working CXX compiler: /usr/bin/c++ -- works
-- Detecting CXX compiler ABI info
-- Detecting CXX compiler ABI info - done
-- Detecting CXX compile features
-- Detecting CXX compile features - done
-- Boost version: 1.65.1
-- Found the following Boost libraries:
--   unit_test_framework
-- Configuring done
CMake Error: Cannot determine link language for target "image".
CMake Error: CMake can not determine linker language for target:
image
-- Generating done
-- Build files have been written to:
```

Specifying project language didn't help
https://stackoverflow.com/a/11802004/1305036

From CMake mailing list I've learnt `set_target_properties` command
https://cmake.org/pipermail/cmake/2006-July/010380.html
I should set `linker_language` property for `image` library to
`CXX`.
So easy.
As I understand,
the point is that `image.hpp` file is not a source file,
and CMake is not quite sure that it's a C++ file.
I don't know for sure.

Next, I should specify include directories of the target.
It was already described in commit
3b2fee105da8950fa38216b3257737a06825dd37
I'll add one note:
if I remove `${CMAKE_CURRENT_SOURCE_DIR}`,
the project will fail on attempt to include `image.hpp` to a module.