Fix llvm-include-order
Another error is a wrong order of includes in test file
https://travis-ci.org/char-lie/stereo-parallel/jobs/443290526

```
/home/travis/build/char-lie/stereo-parallel/tests/image.cpp:24:1: error: #includes are not sorted properly [llvm-include-order,-warnings-as-errors]
^        ~~~~~~~~~~~~~~~~~~~~~~~~~~
         "image.hpp"
```

Documentation to this error refers us to coding standards
https://clang.llvm.org/extra/clang-tidy/checks/llvm-include-order.html
It says that main module header should come first
http://llvm.org/docs/CodingStandards.html#include-style
This clarifies a lot: the test file is called `image.cpp`
and it includes `image.hpp` module in double quotes.
So, the linter recognizes the file as a header of the test.
Use angle brackets to destroy this error.