Add documentation to type aliases in image module
When I've added documentation to type aliases,
it didn't appear in built documentation.
The solution is simple:
I should document the file
https://stackoverflow.com/a/12041646/1305036
This means that I can simply add a comment with `@file` in it.

Now on page of the file in documentation I have a hyperlink
`Go to the documentation of this file.`.