Use 'std::vector' instead of 'std::shared_ptr'
It's more convenient to store data array of `Image` as a `vector`
in order to access its elements and ease its construction.

Looks like Boost Test libraty has no specific tool
for comparison of arrays,
so I'll use a regular `BOOST_CHECK`.
Attempt to use `BOOST_CHECK_EQUAL` led to an error,
because in the case of failure the library cannot display wrong array

```
/usr/include/boost/test/tools/detail/print_helper.hpp:54:18: error: invalid operands to binary expression ('std::ostream' (aka 'basic_ostream<char>') and 'const std::vector<unsigned long, std::allocator<unsigned long> >')
            ostr << t;
            ~~~~ ^  ~
```

There is an alternative called `BOOST_TEST`,
but I cannot get the difference between it and `BOOST_CHECK`,
so I will use the last one,
because it's more consistent with `BOOST_CHECK_EQUAL`.