Implement PGM format reader
Plain PGM image format was chosen,
because it's simple and represents grayscale images
with configurable maximal intensity.
This allows to store disparity map
with values that overflow `255`.

Overloaded `<<` and `>>` operators allow us to read and write images
easily in usual way.
Only `operator>>` needs to be a `friend` of the `PGM_IO` class,
because it uses `read_ppm_instruction` class member
that was decided to be private.