Avoid uninitialized members in Image
The following error wasn't clear enough for me first
https://travis-ci.org/char-lie/stereo-parallel/jobs/457129459

```
/home/travis/build/char-lie/stereo-parallel/include/image.hpp:46:8: error: constructor does not initialize these fields: width, height, max_value [cppcoreguidelines-pro-type-member-init,-warnings-as-errors]
struct Image
       ^
```

Documentation didn't clarify a lot.
I've got that it happened because I don't have a constructor,
but it didn't fail before
https://clang.llvm.org/extra/clang-tidy/checks/cppcoreguidelines-pro-type-member-init.html

If I don't use default constructor and assign a value immediately,
this works fine.
Let it be a solution at the moment.