Implement static member class to check PGM fie end
Here I've found a new fact (for me) about C++
https://softwareengineering.stackexchange.com/a/318091

A C++ application cannot tell you
whether the end of a stream has been reached.
Instead, it can try to read something after the end
and set both `failbit` (it fails to read nothing)
and `eof` (the stream finished).
I was wondering about how to distinguish a fail and end of file
(for example, if the last symbol is wrong),
and a simple end of file.
I've come to a conclusion that `in.eof() || in.good()` check is enough.

I was assured even more when found a note in `istream::get`
documentation.
It says that `failbit` flag is set in the same time
when `EOF` is returned
http://www.cplusplus.com/reference/istream/istream/get/

Okay,
I'll just read the file, ignoring comments
and check white spaces with `std::isspace` function
https://ru.cppreference.com/w/cpp/string/byte/isspace