Implement unit tests to check penalties
It was an interesting trip.
I've had a function `disparity2lowest`.
I've created a LowestPenalties instance using it.
Then, I've changed reparametrization in `DisparityGraph`.
I was wondering why it doesn't affect lowest penalties.
I've changed signature of the function
from `struct LowestPenalties` to `struct LowestPenalties&&`
and returned `std::move(LowestPenalties{graph, pixels, neighborhoods})`.
It "worked".
I guess, it was working because of garbage read,
It should not have worked, because I calculate minimal values in the function,
and any further changes of the graph should not affect it.
So, I've decided to create an ordinary constructor:
it won't be used in parallel code, so it may be created.
Maybe, I should think about creation of some sort of a frozen graph.
It will consume a lot of memory on real images,
but it may help with such curious situations,
and I'll have an ability to provide the frozen graph as a parameter.