Version 0.1
`PROJECT` command of CMake
allows specifying of the project's version.
The syntax is `PROJECT(name VERSION version)`
https://cmake.org/cmake/help/v3.0/command/project.html

This version has a lot of functionality,
and took a lot of time to make.
That's why it's called `0.1`, not `0.0.1`.
The next step is to add parallelization of the CSP procedure.
After this, the version `0.3` should have optimization algorithms
(both CPU and GPU versions).

Let's go!