Implement parallelizable CSP iteration
Each node can be processed independently:
if multiple different threads decided
that an existent node needs to be removed,
and will remove it simultaneously,
there's almost nothing bad in this case.
The only bad thing is
that these threads will do the same work,
and this will eat processor time to no purpose.