Add ability to torn on OpenMP parallelism
Use the same technique as in
c6f7fd797a0b57678fa4db124a0277c0f4d9463d
Add `WITH_OPENMP` flag
(`OFF` by default to not change the default behavior)
and search for the OpenMP library only if the flag is `ON`.
Thus, CMake will look for the library only when the user wants it.
If the library is not found, an error occurs,
because the OpenMP is required in this case.