Mask out the CPP code with `__OPENCL_C_VERSION__`
The `__OPENCL_C_VERSION__` macro contains the version reflecting
the OpenCL version specified during the OpenCL code compilation.
If it's blank, the code is not compiled with OpenCL,
and therefore is an ordinary host (CPU) code.
https://www.khronos.org/registry/OpenCL/sdk/1.2/docs/man/xhtml/preprocessorDirectives.html

This allows us to use C++ constructions where it's possible
(like constructors and namespaces for code readability),
and explicitly specify the cross-device code.