Add OpenCL versions of constants and types
We cannot use `typedef`, `using` neither `const`.
There is a `__constant` keyword,
but it describes another entity as I understand:
it doesn't look like I can define them outside of functions.

We need to use `#define` for all these things.
Why not using `#define` in our `C++` code to not have this mess?
I like the documentation with such explicit things more:
we have types and we distinguish between type and variable definition.

Let's turn back to `BOOL`.
It's presented in OpenCL,
but we are unable to use array of `bool`.
That's why we have to use `int` instead in GPU code.
https://community.khronos.org/t/bool-array-in-opencl/3619
Here is the official documentation on this
ho://www.khronos.org/registry/OpenCL/sdk/1.2/docs/man/xhtml/restrictions.html