Implement `csp_process_pixel` function
The function is convenient to use separately in GPU code,
because the `csp_solution_iteration` uses heuristic
to finish faster if one pixel is completely removed.
However, this slows the GPU code down as a branching,
so it's better to omit this thing there.
Though, the node processing (check + removal) function
is universal for both cases.