Add CUDA program variables initialization
You should use `cudaMalloc` to allocate the GPU memory
(just like `malloc`),
and then use `cudaMemcpy` (like `memcpy`)
to load contents of local arrays into the GPU memory.
In the end, we should copy the result back
and use `cudaFree` to free the memory (just like `free`).

`cudaMemcpy` is different from `memcpy`
in the flag that should be provided:
`cudaMemcpyHostToDevice` if you want to copy from the host to GPU
and `cudaMemcpyDeviceToHost` in the reverse case.

Use the guide to know more
https://devblogs.nvidia.com/how-optimize-data-transfers-cuda-cc/