Implement CUDA CSP solution iteration
The function differs from the corresponding in OpenCL code:
it uses native synchronization ability (available from CUDA 9),
allowing to synchronize not only in current work group of threads,
but all threads (don't know whether it's limited to a kernel or not).
In my tests it appeared that this is faster than synchronization using
an additional global variable.

See example of use
https://stackoverflow.com/a/6405432/1305036
and detailed official documentation
https://devblogs.nvidia.com/cooperative-groups
to know more.