Require CMake 3.11+ and CUDA if the CUDA is needed
CMake 3.11 is the earliest version supporting CUDA language natively
through `enable_language` and `project` commands
https://cmake.org/cmake/help/v3.11/command/enable_language.html
https://cmake.org/cmake/help/v3.11/command/project.html

So, if the CUDA is needed,
we need to require newer version of CMake (not 3.9, but 3.11).
It's a bit problematic for Ubuntu 18.04,
because its default version is 3.10.2
https://launchpad.net/ubuntu/bionic/+source/cmake,
because you should download binaries of newer CMake
from the official assets
https://github.com/Kitware/CMake/releases/download/v3.11.4/cmake-3.11.4-Linux-x86_64.tar.gz
Though, with Ubuntu 20.04 we have CMake 3.16.3
https://launchpad.net/ubuntu/focal/+source/cmake
So, let's wait for the official support of Ubuntu 20.04 from Nvidia.